bridge/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.3.0
httpx>=0.27.0
graphrag>=0.6.0
PyJWT[crypto]>=2.9.0
boto3>=1.34.0
